fix: final iteration
